+++ NEWS.md
@@ -1,3 +1,10 @@
+# Version 5.0.0 (2023-01-27)
+
+- Now there is a 'Rcpp' implementation of the polynomials: functions 
+`SchurPolCPP`, `JackPolCPP`, `ZonalPolCPP` and `ZonalQPolCPP`. It is faster 
+than the Julia implementation.
+
+
 # Version 4.0.0 (2022-12-19)
 
 - The package does not longer depend on the 'gmpoly' package. This dependency